pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "google-cloud-translate>=3.20.1",
    "notion-client>=2.3.0",
    "python-dotenv>=1.0.1",
    "sift-stack-py>=0.4.1",
]
